requirements.txt
sentence-transformers
scikit-learn
requests
regex
chromadb
mysql-connector-python
